Add files via upload
The Simple Calculator project is a basic web application built using HTML, CSS, and JavaScript. The user interface is designed using HTML for structure, with input fields for numbers and buttons for operations like addition, subtraction, multiplication, and division. CSS is used to style the calculator, making it visually clean and user-friendly. The core functionality is implemented using JavaScript, handling the calculations and displaying the results in real time as the user inputs values. This project showcases the integration of front-end web technologies to create an interactive and functional tool.